refactor(liferay): extract portal/ module for site resolution
- Move SiteResolutionPipeline and step factories from inventory/liferay-site-resolver.ts to portal/site-resolver.ts
- Move resolveSite, enrichWithCompanyId, buildSiteChain, fetchGroupInfo, normalizeFriendlyUrl, normalizeLocalizedName, resolvedSiteCache to portal/site-resolution.ts
- Move resolveSiteToken/siteTokenToFriendlyUrl to portal/site-token.ts
- Move getOperationPolicy/OperationName/TransportSurface to portal/capabilities.ts
- Delete inventory/liferay-site-resolver.ts and inventory/capabilities.ts
- Update all callers in inventory/, resource/, content/ and liferay-audit.ts to import directly from portal/
- Unify InventoryDependencies and ResourceDependencies as aliases of SiteResolutionDependencies
- Update test files to point to new portal/ paths
