+++ src/features/liferay/liferay-audit.ts
@@ -1,7 +1,8 @@
 import type {AppConfig} from '../../core/config/load-config.js';
 import {createOAuthTokenClient, type OAuthTokenClient} from '../../core/http/auth.js';
 import {createLiferayApiClient, type HttpApiClient} from '../../core/http/client.js';
-import {fetchPagedItems, resolveSite} from './inventory/liferay-inventory-shared.js';
+import {fetchPagedItems} from './inventory/liferay-inventory-shared.js';
+import {resolveSite} from './portal/site-resolution.js';
 import {createLiferayGateway} from './liferay-gateway.js';
 import {performLiferayHealthCheck} from './liferay-health.js';
