+++ src/features/liferay/resource/liferay-resource-sync-structure.ts
@@ -1,6 +1,6 @@
 import type {AppConfig} from '../../../core/config/load-config.js';
 import type {Printer} from '../../../core/output/printer.js';
-import {resolveSite} from '../inventory/liferay-inventory-shared.js';
+import {resolveSite} from '../portal/site-resolution.js';
 import {resolveStructureFile} from './liferay-resource-paths.js';
 import type {MigrationStats} from './migration/index.js';
 import {syncArtifactDetailed} from './sync-engine.js';
