+++ tests/unit/sync-strategies/fragment-entry-sync-strategy.test.ts
@@ -2,7 +2,7 @@ import {describe, expect, test, vi} from 'vitest';
 
 import type {AppConfig} from '../../../src/core/config/load-config.js';
 import {fragmentEntrySyncStrategy} from '../../../src/features/liferay/resource/sync-strategies/fragment-entry-sync-strategy.js';
-import type {ResolvedSite} from '../../../src/features/liferay/inventory/liferay-site-resolver.js';
+import type {ResolvedSite} from '../../../src/features/liferay/portal/site-resolution.js';
 import type {LocalFragment} from '../../../src/features/liferay/resource/liferay-resource-sync-fragments-types.js';
 import {mockApiClient, mockTokenClient} from './sync-strategy-test-helpers.js';
