+++ tests/unit/sync-strategies/template-sync-strategy.test.ts
@@ -4,7 +4,7 @@ import {describe, expect, test, afterEach} from 'vitest';
 
 import type {AppConfig} from '../../../src/core/config/load-config.js';
 import {templateSyncStrategy} from '../../../src/features/liferay/resource/sync-strategies/template-sync-strategy.js';
-import type {ResolvedSite} from '../../../src/features/liferay/inventory/liferay-site-resolver.js';
+import type {ResolvedSite} from '../../../src/features/liferay/portal/site-resolution.js';
 import {createTempDir} from '../../../src/testing/temp-repo.js';
 
 const mockConfig: AppConfig = {
